# projects/points-web-app/docs/v0.2/points-markets.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Freeism Points–Markets API",
    "version": "0.2.0",
    "description": "Points-owned canonical contract for the twelve APIs consumed by Markets. Better Auth discovery, authorization, token, introspection and revocation endpoints remain standard provider endpoints and are intentionally not duplicated as application operations."
  },
  "servers": [
    {
      "url": "https://points.freeism.app"
    }
  ],
...
  "paths": {
...
    "/api/v1/point-package-auction-eligibility-checks": {
      "post": {
        "operationId": "checkPointPackageAuctionEligibility",
        "summary": "Check Point Package auction eligibility atomically",
        "description": "Checks all items in one Points D1 transaction. One ineligible item produces zero receipts and a safe sorted error list.",
        "security": [
          {
            "m2mOAuth": ["points.packages.auction-eligibility"]
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "description": "Opaque idempotency key. The same key and canonical payload replay the initial HTTP status and domain result.",
            "schema": {
              "$ref": "#/components/schemas/OpaqueId"
            }
          },
          {
            "name": "X-Request-Id",
            "in": "header",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/OpaqueId"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuctionEligibilityRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "headers": {
              "Cache-Control": {
                "description": "Exact cache policy: private, no-store",
                "required": true,
                "schema": {
                  "type": "string",
                  "const": "private, no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuctionEligibilityResponse"
                }
              }
            }
          },
          "409": {
            "description": "Auction ineligibility or idempotency-key reuse conflict. No eligibility receipt is issued.",
            "headers": {
              "Cache-Control": {
                "description": "Exact cache policy: private, no-store",
                "required": true,
                "schema": {
                  "type": "string",
                  "const": "private, no-store"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/AuctionEligibilityProblem"
                    },
                    {
                      "$ref": "#/components/schemas/IdempotencyKeyReusedProblem"
                    }
                  ]
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "default": {
            "description": "RFC 9457 Problem Details",
            "headers": {
              "Cache-Control": {
                "description": "Exact cache policy: private, no-store",
                "required": true,
                "schema": {
                  "type": "string",
                  "const": "private, no-store"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "x-freeism-request-body-max-bytes": 1048576,
        "x-freeism-idempotency": {
          "replayHttpStatus": "preserve-initial",
          "replayDomainResult": true,
          "requestIdMayBeRegenerated": true,
          "sameKeyDifferentPayloadStatus": 409,
          "sameKeyDifferentPayloadCode": "IDEMPOTENCY_KEY_REUSED"
        },
        "x-freeism-cache-control": "private, no-store",
        "x-freeism-receipt": {
          "leaseSeconds": 30,
          "validUntilExclusive": true,
          "boundTo": [
            "clientId",
            "auctionCommandId",
            "auctionCommandHash",
            "allItems.auctionItemId",
            "allItems.pointPackageId",
            "allItems.pointPackageRevisionId",
            "allItems.contentHash"
          ],
          "allOrNothing": true,
          "failureReceiptCount": 0,
          "idempotentReplayExtendsLease": false
        }
      }
    },
    "/api/v1/oauth/link-attempts": {
      "post": {
        "operationId": "createPointsLinkAttempt",
        "summary": "Create a client-authenticated OAuth link attempt",
        "security": [
          {
            "m2mOAuth": ["points.connection.link-attempt.create"]
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "description": "Opaque idempotency key. The same key and canonical payload replay the initial HTTP status and domain result.",
            "schema": {
              "$ref": "#/components/schemas/OpaqueId"
            }
          },
          {
            "name": "X-Request-Id",
            "in": "header",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/OpaqueId"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLinkAttemptRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "headers": {
              "Cache-Control": {
                "description": "Exact cache policy: private, no-store",
                "required": true,
                "schema": {
                  "type": "string",
                  "const": "private, no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateLinkAttemptResponse"
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "default": {
            "description": "RFC 9457 Problem Details",
            "headers": {
              "Cache-Control": {
                "description": "Exact cache policy: private, no-store",
                "required": true,
                "schema": {
                  "type": "string",
                  "const": "private, no-store"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "x-freeism-request-body-max-bytes": 65536,
        "x-freeism-idempotency": {
          "replayHttpStatus": "preserve-initial",
          "replayDomainResult": true,
          "requestIdMayBeRegenerated": true,
          "sameKeyDifferentPayloadStatus": 409,
          "sameKeyDifferentPayloadCode": "IDEMPOTENCY_KEY_REUSED"
        },
        "x-freeism-cache-control": "private, no-store",
        "x-freeism-link-attempt": {
          "maxTtlSeconds": 600,
          "boundTo": [
            "clientId",
            "marketsUserId",
            "stateHash",
            "pkceChallenge",
            "redirectUri",
            "requestedScopes",
            "expiresAt",
            "returnUrlHash"
          ],
          "opaqueAttemptId": true
        }
      }
    },
    "/api/v1/oauth/link-attempts/{linkAttemptId}/finalizations": {
      "post": {
        "operationId": "finalizePointsLinkAttempt",
        "summary": "Confirm or cancel a pending OAuth link attempt",
        "security": [
          {
            "m2mOAuth": ["points.connection.link-attempt.finalize"]
          }
        ],
        "parameters": [
          {
            "name": "linkAttemptId",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/OpaqueId"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "description": "Opaque idempotency key. The same key and canonical payload replay the initial HTTP status and domain result.",
            "schema": {
              "$ref": "#/components/schemas/OpaqueId"
            }
          },
          {
            "name": "X-Request-Id",
            "in": "header",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/OpaqueId"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FinalizeLinkAttemptRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "headers": {
              "Cache-Control": {
                "description": "Exact cache policy: private, no-store",
                "required": true,
                "schema": {
                  "type": "string",
                  "const": "private, no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FinalizeLinkAttemptResponse"
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "default": {
            "description": "RFC 9457 Problem Details",
            "headers": {
              "Cache-Control": {
                "description": "Exact cache policy: private, no-store",
                "required": true,
                "schema": {
                  "type": "string",
                  "const": "private, no-store"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "x-freeism-request-body-max-bytes": 65536,
        "x-freeism-idempotency": {
          "replayHttpStatus": "preserve-initial",
          "replayDomainResult": true,
          "requestIdMayBeRegenerated": true,
          "sameKeyDifferentPayloadStatus": 409,
          "sameKeyDifferentPayloadCode": "IDEMPOTENCY_KEY_REUSED"
        },
        "x-freeism-cache-control": "private, no-store",
        "x-freeism-link-finalization": {
          "tokenExchangeGrantStatus": "PENDING_MARKETS_CONFIRMATION",
          "confirmGrantStatus": "ACTIVE",
          "cancelGrantStatus": "CANCELLED",
          "confirmAfterMarketsLocalPendingSave": true,
          "cancelOrTtlRevokesOnlyNewAttemptGrant": true,
          "sameOutcomeReturnsSameReceipt": true,
          "oppositeOutcomeStatus": 409
        }
      }
    },
    "/api/v1/me/connection": {
      "get": {
        "operationId": "getPointsConnection",
        "summary": "Get the active user's Points connection",
        "security": [
          {
            "userOAuth": ["points.connection.read"]
          }
        ],
        "parameters": [
          {
            "name": "X-Request-Id",
            "in": "header",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/OpaqueId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "headers": {
              "Cache-Control": {
                "description": "Exact cache policy: private, no-store",
                "required": true,
                "schema": {
                  "type": "string",
                  "const": "private, no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PointsConnectionResponse"
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "default": {
            "description": "RFC 9457 Problem Details",
            "headers": {
              "Cache-Control": {
                "description": "Exact cache policy: private, no-store",
                "required": true,
                "schema": {
                  "type": "string",
                  "const": "private, no-store"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "x-freeism-cache-control": "private, no-store"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "userOAuth": {
        "type": "oauth2",
        "description": "JWT Authorization Code + PKCE access tokens for ordinary user operations. Refresh tokens may renew only this user scope set.",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://points.freeism.app/api/auth/oauth2/authorize",
            "tokenUrl": "https://points.freeism.app/api/auth/oauth2/token",
            "refreshUrl": "https://points.freeism.app/api/auth/oauth2/token",
            "scopes": {
              "openid": "OpenID Connect identity",
              "profile": "Basic profile",
              "offline_access": "Refresh token access",
              "points.connection.read": "Read the active Points connection",
              "points.balance.read": "Check point balances",
              "points.reservations.create": "Create point reservations"
            },
            "x-freeism-pkce": {
              "required": true,
              "method": "S256"
            }
          }
        },
        "x-freeism-access-token-format": "jwt",
        "x-freeism-principal-class": "user"
      },
      "connectionUnlinkOAuth": {
        "type": "oauth2",
        "description": "One-time Authorization Code + PKCE token after Google fresh authentication for normal unlink only.",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://points.freeism.app/api/auth/oauth2/authorize",
            "tokenUrl": "https://points.freeism.app/api/auth/oauth2/token",
            "scopes": {
              "points.connection.unlink": "Deactivate exactly one active Points connection"
            },
            "x-freeism-pkce": {
              "required": true,
              "method": "S256"
            }
          }
...
    "schemas": {
      "OpaqueId": {
        "type": "string",
        "minLength": 1,
        "maxLength": 255
      },
...
      "Sha256Hash": {
        "type": "string",
        "pattern": "^sha256:[0-9a-f]{64}$"
      },
...
      "AuctionEligibilityRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "auctionCommandId": {
            "$ref": "#/components/schemas/OpaqueId"
          },
          "auctionCommandHash": {
            "$ref": "#/components/schemas/Sha256Hash"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuctionEligibilityRequestItem"
            },
            "minItems": 1,
            "maxItems": 1000,
            "x-freeism-unique-by": "auctionItemId"
          }
        },
        "required": ["auctionCommandId", "auctionCommandHash", "items"]
      },
...
    }
  }
}